feat(cm): add publish and save
