+++ packages/core/content-manager/shared/contracts/single-types.ts
@@ -1,8 +1,15 @@
-import { EntityService, Common } from '@strapi/types';
+import { Documents, Common } from '@strapi/types';
 
 import { errors } from '@strapi/utils';
 
-type Entity = EntityService.Result<Common.UID.Schema>;
+type Document = Documents.Document<any>;
+type AT_FIELDS = 'updatedAt' | 'createdAt' | 'publishedAt';
+type DocumentMetadata = {
+  // All status of the returned locale
+  availableStatus: Pick<Document, 'id' | 'status' | AT_FIELDS>[];
+  // Available locales within the same status of the returned document
+  availableLocales: Pick<Document, 'id' | 'locale' | 'status' | AT_FIELDS>[];
+};
 
 /**
  * GET /single-types/:model
@@ -19,11 +26,11 @@ export declare namespace Find {
     model: string;
   }
 
-  export type Response =
-    | Entity
-    | {
-        error?: errors.ApplicationError;
-      };
+  export interface Response {
+    data: Document;
+    meta: DocumentMetadata;
+    error?: errors.ApplicationError;
+  }
 }
 
 /**
@@ -31,7 +38,7 @@ export declare namespace Find {
  */
 export declare namespace CreateOrUpdate {
   export interface Request {
-    body: Entity;
+    body: Document;
     query: {
       plugins: {
         i18n: {
@@ -41,11 +48,11 @@ export declare namespace CreateOrUpdate {
     };
   }
 
-  export type Response =
-    | Entity
-    | {
-        error?: errors.ApplicationError;
-      };
+  export interface Response {
+    data: Document;
+    meta: DocumentMetadata;
+    error?: errors.ApplicationError;
+  }
 }
 
 /**
@@ -59,11 +66,11 @@ export declare namespace Delete {
     };
   }
 
-  export type Response =
-    | Entity
-    | {
-        error?: errors.ApplicationError;
-      };
+  export interface Response {
+    data: Document;
+    meta: DocumentMetadata;
+    error?: errors.ApplicationError;
+  }
 }
 
 /**
@@ -77,11 +84,11 @@ export declare namespace Publish {
     };
   }
 
-  export type Response =
-    | Entity
-    | {
-        error?: errors.ApplicationError;
-      };
+  export interface Response {
+    data: Document;
+    meta: DocumentMetadata;
+    error?: errors.ApplicationError;
+  }
 }
 
 /**
@@ -94,11 +101,11 @@ export declare namespace UnPublish {
       locale: string;
     };
   }
-  export type Response =
-    | Entity
-    | {
-        error?: errors.ApplicationError;
-      };
+  export interface Response {
+    data: Document;
+    meta: DocumentMetadata;
+    error?: errors.ApplicationError;
+  }
 }
 
 /**
